+++ CHANGELOG.md
@@ -3,9 +3,14 @@
 ### Features
 
 - **operator**: add `map_to` operator.
+- **operator**: add `finalize` operator.
 - **subscription**: Add `SubscriptionGuard::new()` for enabling RAII for existing subscriptions.
 - **subscription**: Add `SubscriptionWrapper::into_inner()`, e.g. if one wants to add the inner subscription to
   a composite subscription.
+
+### Breaking Changes
+
+- **observable**: don't require items/errors to implement `PayloadCopy`, `Clone` is enough now (remove `PayloadCopy`)
 
 ## [0.8.3](https://github.com/rxRust/rxRust/releases/tag/v0.8.2)  (2020-03-26)